fix: Fixed the photo rendering issue on favorite locations screen.
On the favorite locations page, was able to get the click binding to run a function that logs the correct locations title that was clicked.  Was able to bind that function to start off with locations picture hidden, once clicked the function then shows the photo for every location is list.  Need to fix so that only the location that was clicked picture becomes visible.